rename package to de.fva_net.maxs.logger
