+++ src/main/java/de/fva_net/maxs/logger/MessageType.java
@@ -1,4 +1,4 @@
-package io.github.fvanet.maxsloggerjava;
+package de.fva_net.maxs.logger;
 
 /**
  * Enumeration representing message types.
