+++ src/main/java/de/fva_net/maxs/logger/RexsPart.java
@@ -1,4 +1,4 @@
-package io.github.fvanet.maxsloggerjava;
+package de.fva_net.maxs.logger;
 
 /**
  * Represents a part with a REXS ID.
