+++ src/main/java/de/fva_net/maxs/logger/xml/KernelNotifications.java
@@ -1,4 +1,4 @@
-package io.github.fvanet.maxsloggerjava.xml;
+package de.fva_net.maxs.logger.xml;
 
 import jakarta.xml.bind.annotation.*;
 import lombok.Getter;
