+++ src/main/java/de/fva_net/maxs/logger/xml/Notification.java
@@ -1,7 +1,7 @@
-package io.github.fvanet.maxsloggerjava.xml;
+package de.fva_net.maxs.logger.xml;
 
-import io.github.fvanet.maxsloggerjava.MessageType;
-import io.github.fvanet.maxsloggerjava.StandardRoutine;
+import de.fva_net.maxs.logger.MessageType;
+import de.fva_net.maxs.logger.StandardRoutine;
 import jakarta.xml.bind.annotation.*;
 import lombok.Data;
