+++ src/test/java/de/fva_net/maxs/logger/xml/KernelNotificationsTest.java
@@ -1,4 +1,4 @@
-package io.github.fvanet.maxsloggerjava.xml;
+package de.fva_net.maxs.logger.xml;
 
 import org.junit.jupiter.api.Test;
